backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
loguru==0.7.2
pydantic-settings==2.12.0
pydantic
httpx

google-genai
openai
langchain_core
langgraph